// src/app/Enums/operacion-instruccion.ts
export enum OperacionInstruccion {
  LOAD = 'CARGAR',
  ADD = 'SUMAR',
  SUB = 'RESTAR',
  MUL = 'MULTIPLICAR',
  DIV = 'DIVIDIR',
  MOVE = 'MOVER',
  HALT = 'DETENER',
  INC = 'INCREMENTAR',
  DEC = 'DECREMENTAR',
  AND = 'CONJUNCION',
  OR = 'DISYUNCION',
  NOT = 'NEGACION',
  OUT = 'SALIDA',
  IN = 'ENTRADA',
  JMP = 'SALTO',
  JNE = 'SALTO SI NO IGUAL',
  JE = 'SALTO SI ES IGUAL',
  JG = 'SALTO SI ES MAYOR',
  JL = 'SALTO SI ES MENOR',
  JZ = 'SALTO SI ES CERO',
  JNZ = 'SALTO SI NO ES CERO',
  CMP = 'COMPARAR',
  LOAD_MEM = "LOAD_MEM",
  STORE_MEM = "STORE_MEM"
}